LayoutTests/pointer-lock/lock-fail-responses.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script src="resources/pointer-lock-test-harness.js"></script>
</head>
<body>
<script>
    description("Test asynchronous results of calling lock.")
    window.jsTestIsAsync = true;
    shouldBeDefined("window.testRunner");

    todo = [
        function () {
            testRunner.setPointerLockWillFailSynchronously();
            navigator.webkitPointer.lock(document.body,
                function () {
                    testFailed("Lock succeeded unexpectedly.");
                    finishJSTest();
                },
                function () {
                    testPassed("Lock failed (after we called testRunner.setPointerLockWillFailSynchronously)");
                    doNextStepWithUserGesture();
                });
        },
        function () {
            testRunner.setPointerLockWillFailAsynchronously();
            navigator.webkitPointer.lock(document.body,
                function () {
                    testFailed("Lock succeeded unexpectedly.");
                    finishJSTest();
                },
                function () {
                    testPassed("Lock failed (after we called testRunner.setPointerLockWillFailAsynchronously)");
                    doNextStep();
                });
        },
    ];
    doNextStepWithUserGesture();
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
